PlayerBlock is not part of the board state anymore
Previously the PlayerBlock was considered as part of the state of the
board, meaning that modifying the playerblock position or pieces would
modify the board state too.

Now this will change in order to the view to perform a different
rendering for the player block, and to allow rendering the player block
in between two positions of the board (as the original game does)
This will also allow a future rendering optimization by redrawing only
the part of the block, something we could't do easily if the state
of the board was changed every time that the player block moved.

This means that the ColumnsBoard class is not responsible anymore to move
the block, so instead it provides query methods that the caller can use
to check if movement of the block to a given position is a valid movement

The PlayerBlock pieces will be added to the board state when it is
"consolidated", meaning: when you can't move the block down and thus
the piece is settled.